move compare dicts
